requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-groq
langchain-aws
langgraph
python-dotenv
pydantic
requests
duckduckgo-search
langchain-chroma
chromadb
langchain-text-splitters
langchain-huggingface
sentence-transformers
unstructured
vconsoleprint
